requirements.txt
# Core Data Processing
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0

# Data Collection
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.21.0
lxml==5.2.2

# API & Web Scraping
kaggle==1.6.14
python-dotenv==1.0.1
fake-useragent==1.5.1
retry==0.9.2

# Data Visualization
matplotlib==3.9.0
seaborn==0.13.2
plotly==5.22.0
streamlit==1.35.0

# Machine Learning
scikit-learn==1.5.0
xgboost==2.0.3
lightgbm==4.3.0

# Utilities
tqdm==4.66.4
joblib==1.4.2
pyarrow==16.1.0
openpyxl==3.1.2

# Data Quality & Validation
great-expectations==0.18.15
pandera==0.19.3

# Development Tools
jupyter==1.0.0
ipykernel==6.29.4
black==24.4.2
pylint==3.2.2

# Testing
pytest==8.2.1
pytest-cov==5.0.0

# Database (optional)
sqlalchemy==2.0.30
psycopg2-binary==2.9.9